Refactored code in TestSearchField and added isElementPresent method to Base
