+++ src/main/java/com/academy/ui/pages/Base.java
@@ -1,6 +1,8 @@
 package com.academy.ui.pages;
 
+import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.PageFactory;
 
 public abstract class Base {
@@ -17,6 +19,14 @@ public abstract class Base {
             Thread.sleep(millisSeconds);
         } catch (InterruptedException e) {
 
+        }
+    }
+
+    public boolean isElementPresent(WebElement element){
+        try {
+            return element.isDisplayed();
+        } catch (NoSuchElementException e) {
+            return false;
         }
     }
 }
